Feature: Added cli function to module and refactored
