+++ baseball/__main__.py
@@ -1,0 +1,5 @@
+import sys
+
+if __name__ == '__main__':
+    from baseball.cli import main as _main
+    sys.exit(_main())
