+++ baseball/baseball_helpers.py
@@ -1,0 +1,151 @@
+from collections import OrderedDict
+from re import search, sub, findall
+
+AUTOMATIC_BALL_POSITION = (1.0, 1.0)
+
+POSITION_CODE_DICT = {'pitcher': 1,
+                      'catcher': 2,
+                      'first': 3,
+                      'second': 4,
+                      'third': 5,
+                      'shortstop': 6,
+                      'left': 7,
+                      'center': 8,
+                      'right': 9,
+                      'designated': 10,
+                      '1B': 3,
+                      '2B': 4,
+                      '3B': 5,
+                      'SS': 6,
+                      'P': 1,
+                      'C': 2,
+                      'LF': 7,
+                      'CF': 8,
+                      'RF': 9}
+
+ON_BASE_SUMMARY_DICT = {'Single': '1B',
+                        'Double': '2B',
+                        'Triple': '3B',
+                        'Hit By Pitch': 'HBP',
+                        'Home Run': 'HR',
+                        'Walk': 'BB',
+                        'Intent Walk': 'IBB'}
+
+PLAY_CODE_ORDERED_DICT = OrderedDict([
+    ('picks off', 'PO'),
+    ('caught stealing', 'CS'),
+    ('wild pitch', 'WP'),
+    ('passed ball', 'PB'),
+    ('balk', 'BLK'),
+    ('steals', 'S'),
+    ('fan interference', 'FI'),
+    ('catcher interference', 'CI'),
+    ('error', 'E'),
+    ('ground', 'G'),
+    ('grand slam', 'HR'),
+    ('homers', 'HR'),
+    ('pop', 'P'),
+    ('line', 'L'),
+    ('fly', 'F'),
+    ('flies', 'F'),
+    ('sacrifice fly', 'SF'),
+    ('hit by pitch', 'HBP'),
+    ('bunt', 'B'),
+    ('sacrifice bunt', 'SH'),
+    ('walks', 'BB'),
+    ('intentionally walks', 'IBB'),
+    ('called out on strikes', 'ꓘ'),
+    ('strikes out', 'K'),
+    ('choice', 'FC')
+])
+
+NO_HIT_CODE_LIST = ['K', 'ꓘ', 'BB', 'IBB']
+
+BASE_PLUS_ONE_DICT = {'': '1st',
+                      '1B': '2nd',
+                      '2B': '3rd',
+                      '3B': 'home'}
+
+INCREMENT_BASE_DICT = {'1st': '2nd',
+                       '2nd': '3rd',
+                       '3rd': 'home'}
+
+STADIUM_TIMEZONE_DICT = {
+    'Fenway Park': 'America/New_York',
+    'George M. Steinbrenner Field': 'America/New_York',
+    'Yankee Stadium': 'America/New_York',
+    'Roger Dean Stadium': 'America/New_York',
+    'Joker Marchant Stadium': 'America/New_York',
+    'JetBlue Park': 'America/New_York',
+    'Citi Field': 'America/New_York',
+    'LECOM Park': 'America/New_York',
+    'First Data Field': 'America/New_York',
+    'The Ballpark of the Palm Beaches': 'America/New_York',
+    'Citizens Bank Park': 'America/New_York',
+    'Spectrum Field': 'America/New_York',
+    'Oriole Park': 'America/New_York',
+    'Nationals Park': 'America/New_York',
+    'Champion Stadium': 'America/New_York',
+    'Truist Park': 'America/New_York',
+    'Tropicana Field': 'America/New_York',
+    'Marlins Park': 'America/New_York',
+    'Rogers Centre': 'America/New_York',
+    'PNC Park': 'America/New_York',
+    'Progressive Field': 'America/New_York',
+    'Comerica Park': 'America/New_York',
+    'Great American Ball Park': 'America/New_York',
+    'Miller Park': 'America/Chicago',
+    'Wrigley Field': 'America/Chicago',
+    'Guaranteed Rate Field': 'America/Chicago',
+    'Busch Stadium': 'America/Chicago',
+    'Target Field': 'America/Chicago',
+    'Globe Life Field': 'America/Chicago',
+    'Minute Maid Park': 'America/Chicago',
+    'Kauffman Stadium': 'America/Chicago',
+    'American Family Field': 'America/Chicago',
+    'Coors Field': 'America/Denver',
+    'Chase Field': 'America/Denver',
+    'Safeco Field': 'America/Los_Angeles',
+    'AT&T Park': 'America/Los_Angeles',
+    'Oakland-Alameda County Coliseum': 'America/Los_Angeles',
+    'Oakland Coliseum': 'America/Los_Angeles',
+    'Angel Stadium': 'America/Los_Angeles',
+    'Oracle Park': 'America/Los_Angeles',
+    'Dodger Stadium': 'America/Los_Angeles',
+    'Petco Park': 'America/Los_Angeles',
+    'T-Mobile Park': 'America/Los_Angeles'
+}
+
+EASTERN_TIMEZONE_STR = 'America/New_York'
+
+def strip_this_suffix(pattern, suffix, input_str):
+    match = search(pattern, input_str)
+    while match:
+        start = match.start()
+        end = match.end()
+        str_beginning = input_str[:start]
+        str_middle = sub(suffix, '.', input_str[start:end])
+        str_end = input_str[end:]
+        input_str = str_beginning + str_middle + str_end
+        match = search(pattern, input_str)
+
+    input_str = sub(suffix, '', input_str)
+
+    return input_str.strip()
+
+def strip_suffixes(input_str):
+    input_str = strip_this_suffix(r' Jr\.\s+[A-Z]', r' Jr\.', input_str)
+    input_str = strip_this_suffix(r' Sr\.\s+[A-Z]', r' Sr\.', input_str)
+    input_str = sub(r' II', '', input_str)
+    input_str = sub(r' III', '', input_str)
+    input_str = sub(r' IV', '', input_str)
+    input_str = sub(r' St\. ', ' St ', input_str)
+
+    initials_match = findall(r'([A-Z]\.[A-Z]\.? )', input_str)
+    while initials_match:
+        new_initials = initials_match[0].replace('.', '')
+        input_str = sub(initials_match[0], new_initials, input_str, 1)
+        initials_match = findall(r'([A-Z]\.[A-Z]\.? )', input_str)
+
+    return input_str
+
