+++ baseball/process_game_xml.py
@@ -1,23 +1,15 @@
+
 from datetime import datetime
 from re import search, sub
 
 from dateutil.parser import parse
 from pytz import UTC
 
-from baseball.baseball import (POSITION_CODE_DICT,
-                               PlateAppearance,
-                               Player,
-                               PlayerAppearance,
-                               Inning,
-                               Team,
-                               Game)
-
-from baseball.baseball_events import (AUTOMATIC_BALL_POSITION,
-                                      Pitch,
-                                      Pickoff,
-                                      RunnerAdvance,
-                                      Substitution,
-                                      Switch)
+from baseball import (PlayerAppearance,
+                      Substitution,
+                      Switch,
+                      AUTOMATIC_BALL_POSITION,
+                      POSITION_CODE_DICT)
 
 MLB_TEAM_CODE_DICT = {'LAA': 'ana',
                       'SEA': 'sea',
