+++ setup.py
@@ -1,8 +1,12 @@
 from setuptools import setup
 
+with open('README.md') as f:
+    readme = f.read()
+
 setup(name='baseball',
-      version='16.16',
+      version='16.17',
       description='Library to download, anaylyze, and visualize events in Major League Baseball games.',
+      long_description=readme,
       url='https://github.com/benjamincrom/baseball',
       author='Benjamin B. Crom',
       author_email='benjamincrom@gmail.com',
